requirements.txt
mysql-connector-python==8.3.0
telegram==0.0.1
python-telegram-bot[job-queue]==20.7
python-dateutil==2.8.2
pytz==2023.3.post1
tzwhere==3.0.3
